Plugins and printing have their own managers, use the plugin set by preferences as default.
